+++ packages/flutter_midi_pro/lib/flutter_midi_pro_method_channel.dart
@@ -64,6 +64,11 @@ class MethodChannelFlutterMidiPro extends FlutterMidiProPlatform {
   }
 
   @override
+  Future<void> setOutputDevice(int deviceId) async {
+    await _channel.invokeMethod('setOutputDevice', {'deviceId': deviceId});
+  }
+
+  @override
   Future<void> unloadSoundfont(int sfId) async {
     await _channel.invokeMethod('unloadSoundfont', {'sfId': sfId});
   }